Build flatpak from Headlamp 0.15.0 Linux tarball
This patch simplifies how the flatpak version of Headlamp is built,
as it just creates the flatpak from the officially pre-compiled
Headlamp Linux tarball instead of building the project.

The reason is that too many times we've run into issues generating
all the modules needed for Headlamp. So this way, the process should
go smoother.